refactor name to 'petutils', add workflow
