+++ .github/workflows/run_pytest.yaml
@@ -1,0 +1,27 @@
+name: run_tests
+on:
+  push:
+  workflow_call:
+    workflow_dispatch:
+    inputs:
+      debug_enabled:
+        type: boolean
+        description: 'Run the build with tmate debugging enabled (https://github.com/marketplace/actions/debugging-with-tmate)'
+        required: false
+        default: false
+
+jobs:
+  test_with_pytest:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - name: Install Poetry
+        run: pipx install poetry
+      - uses: actions/setup-python@v4
+        with:
+          python-version: 3.11
+          cache: poetry
+      - name: Install dependencies
+        run: poetry install
+      - name: Run Tests
+        run: poetry run pytest tests/
